Convert the Base and Pattern editors to inline expansion
The 03 section's color layers are no longer popovers: the two cards are
input-chrome toggles (Mantine Input as button, focus border while open)
sharing one Collapse drawer below the pair, per the variant-C prototype
decision on #460. The editor reuses kit controls: ColorInput for colors,
ListLengthActions for stops (first stop at 0, second at 1, minus removes
the bottom entry); stop reordering and per-row delete are gone, and stop
rows are plain rows rather than nested Surfaces. Per-layer mode memory
(solid/linear/radial) is preserved across card switches and drawer closes.
FactionBackgroundColorLayer.tsx dissolves into the section file; its
non-portalled 440px popover is retired.

Closes #465
