--- src/app/widgets/faction-editor/FactionBackgroundColorLayer.tsx
@@ -1,342 +1,0 @@
-import {
-  Box,
-  Button,
-  ColorInput,
-  Group,
-  NumberInput,
-  Popover,
-  SegmentedControl,
-  SimpleGrid,
-  Stack,
-  Text,
-  UnstyledButton,
-} from '@mantine/core';
-import { ControlBlock } from '@ui/control/ControlBlock';
-import { IconAction } from '@ui/control/IconAction';
-import { Surface } from '@ui/surface';
-import { ArrowDown, ArrowUp, Plus, Trash2 } from 'lucide-react';
-import { useEffect, useId, useRef } from 'react';
-
-import type { Faction } from '@db/factions';
-import { GradientDef } from '@game/assets/utils/Background';
-
-type ColorLayer = Faction['background']['colors'][number];
-type LinearLayer = Extract<ColorLayer, { type: 'linear' }>;
-type RadialLayer = Extract<ColorLayer, { type: 'radial' }>;
-type GradientLayer = LinearLayer | RadialLayer;
-type ColorStop = GradientLayer['stops'][number];
-
-function firstStopColor(value: GradientLayer): string {
-  return value.stops[0]?.[0] ?? '#444444';
-}
-
-function defaultStops(color: string): ColorStop[] {
-  return [
-    [color, 0],
-    [color, 1],
-  ];
-}
-
-function asOptionalNumber(value: string | number): number | undefined {
-  return typeof value === 'number' && Number.isFinite(value) ? value : undefined;
-}
-
-// Same square viewBox + slice crop as the sheet renderer, so the swatch shows
-// the sheet's gradient geometry rather than a CSS reinterpretation of it.
-function LayerSwatch({ value }: { value: ColorLayer }) {
-  const id = `swatch-${useId().replace(/:/g, '')}`;
-  return (
-    <svg
-      viewBox="0 0 100 100"
-      preserveAspectRatio="xMidYMid slice"
-      aria-hidden
-      style={{ display: 'block', width: '100%', height: '100%' }}
-    >
-      {typeof value === 'string' ? null : (
-        <defs>
-          <GradientDef id={id} gradient={value} />
-        </defs>
-      )}
-      <rect width="100" height="100" fill={typeof value === 'string' ? value : `url(#${id})`} />
-    </svg>
-  );
-}
-
-export function FactionBackgroundColorLayer({
-  label,
-  description,
-  value,
-  onChange,
-}: {
-  label: string;
-  description: string;
-  value: ColorLayer;
-  onChange: (value: ColorLayer) => void;
-}) {
-  const mode = typeof value === 'string' ? 'solid' : value.type;
-  const rememberedSolid = useRef(typeof value === 'string' ? value : firstStopColor(value));
-  const rememberedLinear = useRef<LinearLayer | undefined>(
-    typeof value !== 'string' && value.type === 'linear' ? structuredClone(value) : undefined
-  );
-  const rememberedRadial = useRef<RadialLayer | undefined>(
-    typeof value !== 'string' && value.type === 'radial' ? structuredClone(value) : undefined
-  );
-
-  useEffect(() => {
-    if (typeof value === 'string') {
-      rememberedSolid.current = value;
-    } else if (value.type === 'linear') {
-      rememberedLinear.current = structuredClone(value);
-    } else {
-      rememberedRadial.current = structuredClone(value);
-    }
-  }, [value]);
-
-  const changeMode = (nextMode: string) => {
-    if (nextMode === mode) {
-      return;
-    }
-    const sourceColor = typeof value === 'string' ? value : firstStopColor(value);
-    if (nextMode === 'solid') {
-      onChange(rememberedSolid.current || sourceColor);
-    } else if (nextMode === 'linear') {
-      onChange(
-        rememberedLinear.current ?? {
-          type: 'linear',
-          angle: 90,
-          stops: typeof value === 'string' ? defaultStops(value) : structuredClone(value.stops),
-        }
-      );
-    } else {
-      onChange(
-        rememberedRadial.current ?? {
-          type: 'radial',
-          stops: typeof value === 'string' ? defaultStops(value) : structuredClone(value.stops),
-        }
-      );
-    }
-  };
-
-  return (
-    <Popover width={440} position="bottom-start" shadow="md" withinPortal={false}>
-      <Popover.Target>
-        <UnstyledButton type="button" aria-label={`Edit ${label.toLowerCase()} color layer`}>
-          <Surface padding="sm">
-            <Group justify="space-between" gap="sm" wrap="nowrap">
-              <Box>
-                <Text fw={700}>{label}</Text>
-                <Text size="xs" c="dimmed">
-                  {mode === 'solid' ? 'Solid color' : `${mode} gradient`}
-                </Text>
-              </Box>
-              <Box
-                w={76}
-                h={42}
-                style={{
-                  border: '1px solid var(--mantine-color-gray-4)',
-                  borderRadius: 'var(--mantine-radius-sm)',
-                  overflow: 'hidden',
-                }}
-              >
-                <LayerSwatch value={value} />
-              </Box>
-            </Group>
-          </Surface>
-        </UnstyledButton>
-      </Popover.Target>
-      <Popover.Dropdown>
-        <Stack gap="md">
-          <Stack gap="xs">
-            <Text fw={700}>{label}</Text>
-            <Text size="xs" c="dimmed">
-              {description}
-            </Text>
-          </Stack>
-
-          <SegmentedControl
-            fullWidth
-            value={mode}
-            onChange={changeMode}
-            data={[
-              { value: 'solid', label: 'Solid' },
-              { value: 'linear', label: 'Linear' },
-              { value: 'radial', label: 'Radial' },
-            ]}
-            aria-label={`${label} color mode`}
-          />
-
-          {typeof value === 'string' ? (
-            <ColorInput label={`${label} color`} value={value} format="hex" swatchesPerRow={6} onChange={onChange} />
-          ) : (
-            <GradientFields value={value} onChange={onChange} label={label} />
-          )}
-        </Stack>
-      </Popover.Dropdown>
-    </Popover>
-  );
-}
-
-function GradientFields({
-  value,
-  onChange,
-  label,
-}: {
-  value: GradientLayer;
-  onChange: (value: ColorLayer) => void;
-  label: string;
-}) {
-  const updateStops = (stops: ColorStop[]) => onChange({ ...value, stops } as GradientLayer);
-
-  return (
-    <Stack gap="md">
-      {value.type === 'linear' ? (
-        <ControlBlock
-          title="Gradient angle"
-          description="Direction in degrees. The complete admitted range is 0–360."
-          input={
-            <NumberInput
-              aria-label="Gradient angle"
-              min={0}
-              max={360}
-              step={1}
-              allowDecimal={false}
-              value={value.angle}
-              suffix="°"
-              onChange={(next) =>
-                onChange({
-                  ...value,
-                  angle: typeof next === 'number' && Number.isInteger(next) ? Math.min(360, Math.max(0, next)) : 0,
-                })
-              }
-            />
-          }
-        />
-      ) : (
-        <SimpleGrid cols={{ base: 1, xs: 3 }}>
-          {(['x', 'y', 'r'] as const).map((property) => (
-            <NumberInput
-              key={property}
-              label={
-                property === 'x'
-                  ? 'Center X (optional)'
-                  : property === 'y'
-                    ? 'Center Y (optional)'
-                    : 'Radius (optional)'
-              }
-              placeholder="Renderer default"
-              value={value[property] ?? ''}
-              decimalScale={2}
-              onChange={(next) =>
-                onChange({
-                  ...value,
-                  [property]: asOptionalNumber(next),
-                })
-              }
-            />
-          ))}
-        </SimpleGrid>
-      )}
-
-      <ControlBlock
-        title="Gradient stops"
-        description="Ordered colors with positions from 0 to 1. Existing uncommon arrays remain editable."
-        tool={
-          <Button
-            type="button"
-            variant="light"
-            color="dune"
-            size="compact-sm"
-            leftSection={<Plus size={14} aria-hidden />}
-            onClick={() => {
-              const last = value.stops.at(-1);
-              updateStops([...value.stops, [last?.[0] ?? '#888888', last ? last[1] : 0.5]]);
-            }}
-          >
-            Add stop
-          </Button>
-        }
-        input={
-          <Stack gap="xs">
-            {value.stops.length === 0 ? (
-              <Text size="sm" c="dimmed" fs="italic">
-                This gradient has no stops. Add one to make its color visible.
-              </Text>
-            ) : null}
-
-            {value.stops.map((stop, index) => (
-              <Surface key={`${label}-stop-${index}`} padding="sm">
-                <SimpleGrid cols={{ base: 1, xs: 2 }}>
-                  <ColorInput
-                    label={`Stop ${index + 1} color`}
-                    value={stop[0]}
-                    format="hex"
-                    onChange={(color) => {
-                      const next = [...value.stops];
-                      next[index] = [color, stop[1]];
-                      updateStops(next);
-                    }}
-                  />
-                  <NumberInput
-                    label={`Stop ${index + 1} position`}
-                    min={0}
-                    max={1}
-                    step={0.01}
-                    decimalScale={2}
-                    value={stop[1]}
-                    onChange={(position) => {
-                      const next = [...value.stops];
-                      next[index] = [
-                        stop[0],
-                        typeof position === 'number' ? Math.min(1, Math.max(0, position)) : stop[1],
-                      ];
-                      updateStops(next);
-                    }}
-                  />
-                </SimpleGrid>
-                <Group justify="flex-end" gap={4} mt="xs">
-                  <IconAction
-                    label={`Move stop ${index + 1} earlier`}
-                    variant="subtle"
-                    color="gray"
-                    disabled={index === 0}
-                    onClick={() => {
-                      if (index === 0) {
-                        return;
-                      }
-                      const next = [...value.stops];
-                      [next[index - 1], next[index]] = [next[index], next[index - 1]];
-                      updateStops(next);
-                    }}
-                    icon={<ArrowUp size={16} aria-hidden />}
-                  />
-                  <IconAction
-                    label={`Move stop ${index + 1} later`}
-                    variant="subtle"
-                    color="gray"
-                    disabled={index === value.stops.length - 1}
-                    onClick={() => {
-                      if (index === value.stops.length - 1) {
-                        return;
-                      }
-                      const next = [...value.stops];
-                      [next[index], next[index + 1]] = [next[index + 1], next[index]];
-                      updateStops(next);
-                    }}
-                    icon={<ArrowDown size={16} aria-hidden />}
-                  />
-                  <IconAction
-                    label={`Remove stop ${index + 1}`}
-                    variant="light"
-                    color="red"
-                    onClick={() => updateStops(value.stops.filter((_, i) => i !== index))}
-                    icon={<Trash2 size={16} aria-hidden />}
-                  />
-                </Group>
-              </Surface>
-            ))}
-          </Stack>
-        }
-      />
-    </Stack>
-  );
-}
